relocate haversine and hausdorff distance
