+++ cpp/include/cuspatial/distance/haversine.hpp
