+++ cpp/tests/spatial/hausdorff_test.cpp
@@ -16,7 +16,7 @@
 
 #include <cudf/column/column_view.hpp>
 #include <cuspatial/error.hpp>
-#include <cuspatial/hausdorff.hpp>
+#include <cuspatial/distance/hausdorff.hpp>
 
 #include <cudf_test/base_fixture.hpp>
 #include <cudf_test/column_utilities.hpp>
